+++ R/color_repel.R
@@ -13,6 +13,7 @@
 #' @param layer layer to detect color, defaults to first
 #' @param out_orig output the original colors as named vector
 #' @param out_worst output the worst combination instead of best
+#' @param ggbuild already built ggplot_built object if available
 #' @examples
 #' a <- ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(displ, hwy)) +
 #'   ggplot2::geom_point(ggplot2::aes(color = as.factor(cyl)))
@@ -33,16 +34,21 @@ color_repel <- function(g,
                         autoswitch = TRUE,
                         layer = 1,
                         out_orig = FALSE,
-                        out_worst = FALSE) {
+                        out_worst = FALSE,
+                        ggbuild = NULL) {
   g <- check_patchwork(g)
 
   if (verbose) {
     message("extract original colors...")
   }
-  temp <- check_colour_mapping(g, col = col, return_col = TRUE, autoswitch = autoswitch, layer = layer)
+  temp <- check_colour_mapping(g, col = col, return_col = TRUE, autoswitch = autoswitch, layer = layer, ggbuild = ggbuild)
   col <- temp[["col"]]
   cols <- temp[["cols"]]
-  g2 <- ggplot2::ggplot_build(g)
+  if (is.null(ggbuild)) {
+    g2 <- ggplot2::ggplot_build(g)
+  } else {
+    g2 <- ggbuild
+  }
 
   if (length(cols) <= 1) {
     warning("Did not detect multiple colors, did you specify the correct mapping? Trying to autoswitch...")
@@ -102,9 +108,11 @@ color_repel <- function(g,
       message("extract plot distances (part 2)...")
     }
     rownames(cdist) <- as.character(1:nrow(cdist))
-    cdist <- suppressMessages(average_clusters_rowwise(cdist, metadata = clust, 
-                                                       if_log = FALSE, method = "min", 
-                                                       output_log = FALSE, trim = TRUE))
+    cdist <- suppressMessages(average_clusters_rowwise(cdist,
+      metadata = clust,
+      if_log = FALSE, method = "min",
+      output_log = FALSE, trim = TRUE
+    ))
     ord <- gtools::mixedorder(colnames(cdist))
     cdist <- cdist[ord, ord]
     cdist[cdist < max(cdist) / 100] <- max(cdist) / 100